deployment workflow file rename
